+++ src/lib/data/dark_elf.rb
@@ -5,11 +5,11 @@ module Charred
     def load_dark_elf(data)
       file = File.read('data/dark_elf/skills.json')
       skills = JSON.parse(file)
-      data[:skills].merge! skills
+      verbose_merge data[:skills], skills
 
       file = File.read('data/dark_elf/traits.json')
       traits = JSON.parse(file)
-      data[:traits].merge! traits
+      verbose_merge data[:traits], traits
 
       file = File.read('data/dark_elf/lifepaths.json')
       lifepaths = JSON.parse(file)